Major refactoring & upgrade (#51)
* feat(app): replaced ConnectKit with Privy

* feat(app): added truncatted addess when logged in

* chore(indexer): from npm to yarn

* test(contracts): added mine_blocks.sh script

* chore(contracts): renamed DeployZKube.sol to DeployZKube.s.sol

* test(contracts): updated cube composer puzzle set

* feat(contracts): added getGame function to ZKube.sol

* feat(app): functioning multiplayer version

* style(app): updated game screen

* feat(app): mobile testing

* feat(contracts): added Arbitrum support

* feat(app): updated ABI

* refactor(circuits): error message throwing in exportCalldataGroth16

* reafctor(app): correctly typed useProof hook

* styles(app): fixed LoginCTA styling

* styles(app): altered canvas camera position

* feat(app): added react-device-detect

* fix(app): usePrivyWalletAddress wallet updates only if `walletClientType` === 'privy'

* feat(app): CopyAddressToClipboardButton closes a tooltip after 1500ms

* feat(app): added Arbitrum Sepolia

* reafctor(app): renamed exportCallDataGroth16 to generateGroth16ProofCalldata

* feat(app): added useCurrentRound hook

* chore(app): updated contract addresses

* refactor(app): renamed useGameData to useGameAndPuzzleData & adjusted logic to new contract return type

* refactor(app): moved SinglePlayerGame into page

* refactor(app): update useContract hook

* refactor(app): refactored multiplayer flow

* refactor(app): updated create game flow

* refactor(app): removed GamesTable

* refactor(app): moved modal to separate dir

* style(app): styled Function component

* feat(app): updated game screens

* feat(app): updated usePrivyWalletAddress to update on logout

* fix(app): ?? to ||

* refactor(app): defined a return type for a callback hook

* feat(app): added callback hooks

* refactor(app): refactoring (decoupling) useClient and useContracts

* refactor(app): removed unused compoents GameStarting and getGames

* refactor(app): refactor(app): major refactoring - callback calls, memoisation, decoupling, types...

* refactor(app): moved hooks calling view functions into `fetch` dir

* styles(app): fixed tailwind className ordering with prettier-plugin-tailwindcss

* refactor(app): moved `callback.types.ts` to `types` dir from `hooks` dir

* refactor(app): added types for fetch hooks

* styles(app): added global css classes for buttons

* refactor(app): decouple Function and DisplayName

* refactor(app): LoadingState component decoupled

* fix(app): puzzle CTA to top

* feat(app): added page footer

* fix(app): fixed ContractFetchReturnType import

* refactor(app): removed Apollo

* chore(app): removed GameContext

* fix(app): added margin-top to loading screens

* style(app): global modal styling

* chore: updated readme

* style(app): modal width and height

* style(app): updated game modal

* styles(app): added alias imports

* lint(app): removed log

* chore(app): added env to next.config.js

---------

+++ contracts/test/ZKube.harness.sol
@@ -6,7 +6,7 @@ import {ZKube} from "../src/ZKube.sol";
 contract ZKubeHarness is ZKube {
     constructor(address verifier) ZKube(verifier) {}
 
-    function exposed_getBlock(uint8 interval, uint72 startingBlock, uint16 numberOfRounds)
+    function exposed_getBlock(uint16 interval, uint72 startingBlock, uint8 numberOfRounds)
         public
         view
         returns (uint256 blockNumber)
